move Git Reference Deserialization to separate crate
