+++ forest/Cargo.toml
@@ -6,6 +6,11 @@ description = "Filecoin implementation in Rust. This command will start the daem
 edition = "2021"
 build = "build.rs"
 
+[lib]
+name = "forest"
+path = "src/lib.rs"
+
+
 [dependencies]
 fvm = "0.2"
 address = { package = "forest_address", version = "0.3" }
@@ -73,10 +78,7 @@ rand = "0.8"
 tempfile = "3"
 
 [build-dependencies]
-#git-version = "0.3"
-serde = { version = "1.0", features = ["derive"] }
-serde_derive = "1.0.106"
-toml = "0.5"
+forest_git_utils = { path = "../utils/git_utils" }
 
 [features]
 default = ["rocksdb", "networks/mainnet"]
