+++ node/forest_libp2p/Cargo.toml
@@ -50,7 +50,7 @@ tiny-cid = "0.2"
 ipld_blockstore = "0.1"
 async-trait = "0.1"
 lazy_static = "1.4"
-git-version = "0.3"
+forest_git_utils = { path = "../../utils/git_utils" }
 serde_ipld_dagcbor = "0.1.0"
 
 [dev-dependencies]
