+++ node/forest_libp2p/src/behaviour.rs
@@ -16,9 +16,9 @@ use crate::{
 };
 use forest_cid::Cid;
 use forest_encoding::blake2b_256;
+use forest_git_utils::CURRENT_COMMIT;
 use futures::channel::oneshot::{self, Sender as OneShotSender};
 use futures::{prelude::*, stream::FuturesUnordered};
-use git_version::git_version;
 use libp2p::ping::{Ping, PingEvent};
 use libp2p::request_response::{
     ProtocolSupport, RequestId, RequestResponse, RequestResponseConfig, RequestResponseEvent,
@@ -55,8 +55,8 @@ use std::{task::Context, task::Poll};
 use tiny_cid::Cid as Cid2;
 
 lazy_static! {
-    static ref VERSION: &'static str = env!("FOREST_VERSION");
-    static ref CURRENT_COMMIT: &'static str = env!("CURRENT_COMMIT");
+    static ref VERSION: &'static str = env!("CARGO_PKG_VERSION");
+//    static ref CURRENT_COMMIT: &'static str = forest_git_utils::CURRENT_COMMIT.as_str();
 }
 
 /// Libp2p behaviour for the Forest node. This handles all sub protocols needed for a Filecoin node.
