+++ utils/git_utils/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "forest_git_utils"
+description = "Utilities used to interact with Git or a serialized Commit File"
+license = "MIT OR Apache-2.0"
+version = "0.0.1"
+authors = ["ChainSafe Systems <info@chainsafe.io>"]
+edition = "2021"
+repository = "https://github.com/ChainSafe/forest"
+
+[dependencies]
+#git-version = "0.3"
+lazy_static = "1.4"
+serde = { version = "1.0" }
+serde_derive = "1.0.106"
+toml = "0.5"
+
+[dev-dependencies]
+serde = { version = "1.0", features = ["derive"] }
